reorg and add more comments
