+++ ffi/examples/read-table/read_table.c
@@ -1,8 +1,6 @@
 #include <stdio.h>
 #include <string.h>
 
-#include <delta_kernel_ffi.h>
-
 #include "read_table.h"
 #include "schema.h"
 
@@ -10,24 +8,7 @@
 #include "arrow.h"
 #endif
 
-struct EngineContext {
-  GlobalScanState* global_state;
-  char* table_root;
-  const ExternEngineHandle* engine;
-  PartitionList* partition_cols;
-  CStringMap* partition_values;
-#ifdef PRINT_ARROW_DATA
-  ArrowContext* arrow_context;
-#endif
-};
-
-// This is how we represent our errors. The kernel will ask us to contruct this struct whenever it
-// enounters an error, and then return the contructed EngineError to us
-typedef struct Error {
-  struct EngineError etype;
-  char* msg;
-} Error;
-
+// Print the content of a selection vector if `VERBOSE` is defined in read_table.h
 void print_selection_vector(const char* indent, const KernelBoolSlice* selection_vec) {
 #ifdef VERBOSE
   for (int i = 0; i < selection_vec->len; i++) {
@@ -36,100 +17,7 @@ void print_selection_vector(const char* indent, const KernelBoolSlice* selection
 #endif
 }
 
-EngineError* allocate_error(KernelError etype, const KernelStringSlice msg) {
-  Error* error = malloc(sizeof(Error));
-  error->etype.etype = etype;
-  char* charmsg = allocate_string(msg);
-  error->msg = charmsg;
-  return (EngineError*)error;
-}
-
-void free_error(Error* error) {
-  free(error->msg);
-  free(error);
-}
-
-void print_error(const char* indent, Error* err) {
-  printf("%sCode: %i\n", indent, err->etype);
-  printf("%sMsg: %s\n", indent, err->msg);
-}
-
-void set_builder_opt(EngineBuilder* engine_builder, char* key, char* val) {
-  KernelStringSlice key_slice = { key, strlen(key) };
-  KernelStringSlice val_slice = { val, strlen(val) };
-  set_builder_option(engine_builder, key_slice, val_slice);
-}
-
-#ifdef PRINT_ARROW_DATA
-// convert to a garrow boolean array. can't use garrow_boolean_array_builder_append_values as that
-// expects a gboolean*, which is actually an int* which is 4 bytes, but our slice is a C99 _Bool*
-// which is 1 byte
-GArrowBooleanArray* slice_to_arrow_bool_array(const KernelBoolSlice slice) {
-  GArrowBooleanArrayBuilder* builder = garrow_boolean_array_builder_new();
-  for (int i = 0; i < slice.len; i++) {
-    gboolean val = slice.ptr[i] ? TRUE : FALSE;
-    garrow_boolean_array_builder_append_value(builder, val, NULL);
-  }
-  GArrowArray* ret = garrow_array_builder_finish((GArrowArrayBuilder*)builder, NULL);
-  if (ret == NULL) {
-    printf("ERROR IN BUILDING\n");
-  }
-  return (GArrowBooleanArray*)ret;
-}
-
-void visit_read_data(void* vcontext, EngineDataHandle* data) {
-  print_diag("  Converting read data to arrow and adding to context\n");
-  struct EngineContext* context = vcontext;
-  ExternResultArrowFFIData arrow_res = get_raw_arrow_data(data, context->engine);
-  if (arrow_res.tag != OkArrowFFIData) {
-    printf("Failed to get arrow data\n");
-    print_error("  ", (Error*)arrow_res.err);
-    free_error((Error*)arrow_res.err);
-    exit(-1);
-  }
-  ArrowFFIData* arrow_data = arrow_res.ok;
-  add_batch_to_context(
-    context->arrow_context, arrow_data, context->partition_cols, context->partition_values);
-}
-
-void read_parquet_file(struct EngineContext* context,
-                       const KernelStringSlice path,
-                       const KernelBoolSlice selection_vector) {
-  int full_len = strlen(context->table_root) + path.len + 1;
-  char* full_path = malloc(sizeof(char) * full_len);
-  snprintf(full_path, full_len, "%s%.*s", context->table_root, path.len, path.ptr);
-  print_diag("  Reading parquet file at %s\n", full_path);
-  KernelStringSlice path_slice = { full_path, full_len };
-  Schema* read_schema = get_global_read_schema(context->global_state);
-  FileMeta meta = {
-    .path = path_slice,
-  };
-  ExternResultFileReadResultIterator read_res =
-    read_parquet_files(context->engine, &meta, read_schema);
-  if (read_res.tag != OkFileReadResultIterator) {
-    printf("Couldn't read data\n");
-    return;
-  }
-  if (selection_vector.len > 0) {
-    GArrowBooleanArray* sel_array = slice_to_arrow_bool_array(selection_vector);
-    context->arrow_context->cur_filter = sel_array;
-  }
-  FileReadResultIterator* read_iter = read_res.ok;
-  for (;;) {
-    ExternResultbool ok_res = read_result_next(read_iter, context, visit_read_data);
-    if (ok_res.tag != Okbool) {
-      printf("Failed to iterate read data\n");
-      print_error("  ", (Error*)ok_res.err);
-      free_error((Error*)ok_res.err);
-      exit(-1);
-    } else if (!ok_res.ok) {
-      print_diag("  Done reading parquet file\n");
-      break;
-    }
-  }
-}
-#endif
-
+// Print info about table partitions if `VERBOSE` is defined in read_table.h
 void print_partition_info(struct EngineContext* context, CStringMap* partition_values) {
 #ifdef VERBOSE
   for (int i = 0; i < context->partition_cols->len; i++) {
@@ -146,11 +34,30 @@ void print_partition_info(struct EngineContext* context, CStringMap* partition_v
 #endif
 }
 
-void visit_callback(void* engine_context,
-                    const KernelStringSlice path,
-                    long size,
-                    const DvInfo* dv_info,
-                    CStringMap* partition_values) {
+// kernel will call this to allocate our errors. This can be used to create an "engine native" type
+// error
+EngineError* allocate_error(KernelError etype, const KernelStringSlice msg) {
+  Error* error = malloc(sizeof(Error));
+  error->etype.etype = etype;
+  char* charmsg = allocate_string(msg);
+  error->msg = charmsg;
+  return (EngineError*)error;
+}
+
+// utility function to convert key/val into slices and set them on a builder
+void set_builder_opt(EngineBuilder* engine_builder, char* key, char* val) {
+  KernelStringSlice key_slice = { key, strlen(key) };
+  KernelStringSlice val_slice = { val, strlen(val) };
+  set_builder_option(engine_builder, key_slice, val_slice);
+}
+
+// Kernel will call this function for each file that should be scanned. The arguments include enough
+// context to constuct the correct logical data from the physically read parquet
+void scan_row_callback(void* engine_context,
+			const KernelStringSlice path,
+			long size,
+			const DvInfo* dv_info,
+			CStringMap* partition_values) {
   struct EngineContext* context = engine_context;
   print_diag("Called back to read file: %.*s\n", (int)path.len, path.ptr);
   ExternResultKernelBoolSlice selection_vector_res =
@@ -175,16 +82,22 @@ void visit_callback(void* engine_context,
   context->partition_values = NULL;
 }
 
-void visit_data(void* engine_context,
-                EngineDataHandle* engine_data,
-                KernelBoolSlice selection_vec) {
+// For each chunk of scan data (which may contain multiple files to scan), kernel will call this
+// function (named do_visit_scan_data to avoid conflict with visit_scan_data exported by kernel)
+void do_visit_scan_data(void* engine_context,
+		     EngineDataHandle* engine_data,
+		     KernelBoolSlice selection_vec) {
   print_diag("\nScan iterator found some data to read\n  Of this data, here is "
              "a selection vector\n");
   print_selection_vector("    ", &selection_vec);
-  visit_scan_data(engine_data, selection_vec, engine_context, visit_callback);
+  // Ask kernel to iterate each individual file and call us back with extracted metadata
+  visit_scan_data(engine_data, selection_vec, engine_context, scan_row_callback);
   drop_bool_slice(selection_vec);
 }
 
+// Called for each element of the partition StringSliceIterator. We just turn the slice into a
+// `char*` and append it to our list. We knew the total number of partitions up front, so this
+// assumes that `list->cols` has been allocated with enough space to store the pointer.
 void visit_partition(void* context, const KernelStringSlice partition) {
   PartitionList* list = context;
   char* col = allocate_string(partition);
@@ -192,9 +105,11 @@ void visit_partition(void* context, const KernelStringSlice partition) {
   list->len++;
 }
 
+// Build a list of partition column names.
 PartitionList* get_partition_list(GlobalScanState* state) {
   int count = get_partition_column_count(state);
   PartitionList* list = malloc(sizeof(PartitionList));
+  // We set the `len` to 0 here and use it to track how many items we've added to the list
   list->len = 0;
   list->cols = malloc(sizeof(char*) * count);
   StringSliceIterator* part_iter = get_partition_columns(state);
@@ -204,6 +119,10 @@ PartitionList* get_partition_list(GlobalScanState* state) {
       print_diag("Done iterating partition columns");
       break;
     }
+  }
+  if (list->len != count) {
+    printf("Error, partition iterator did not return get_partition_column_count columns\n");
+    exit(-1);
   }
   return list;
 }
@@ -297,7 +216,7 @@ int main(int argc, char* argv[]) {
 
   // iterate scan files
   for (;;) {
-    ExternResultbool ok_res = kernel_scan_data_next(data_iter, &context, visit_data);
+    ExternResultbool ok_res = kernel_scan_data_next(data_iter, &context, do_visit_scan_data);
     if (ok_res.tag != Okbool) {
       printf("Failed to iterate scan data\n");
       print_error("  ", (Error*)ok_res.err);
